+++ dash_renderer/dash_renderer.dev.js
@@ -33092,11 +33092,9 @@ Object.defineProperty(exports, "__esModule", {
     value: true
 });
 
+var _slicedToArray = function () { function sliceIterator(arr, i) { var _arr = []; var _n = true; var _d = false; var _e = undefined; try { for (var _i = arr[Symbol.iterator](), _s; !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i["return"]) _i["return"](); } finally { if (_d) throw _e; } } return _arr; } return function (arr, i) { if (Array.isArray(arr)) { return arr; } else if (Symbol.iterator in Object(arr)) { return sliceIterator(arr, i); } else { throw new TypeError("Invalid attempt to destructure non-iterable instance"); } }; }();
+
 var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
-
-var _ramda = __webpack_require__(/*! ramda */ "./node_modules/ramda/index.js");
-
-var _ramda2 = _interopRequireDefault(_ramda);
 
 var _react = __webpack_require__(/*! react */ "react");
 
@@ -33113,6 +33111,12 @@ var _registry2 = _interopRequireDefault(_registry);
 var _NotifyObservers = __webpack_require__(/*! ./components/core/NotifyObservers.react */ "./src/components/core/NotifyObservers.react.js");
 
 var _NotifyObservers2 = _interopRequireDefault(_NotifyObservers);
+
+var _reactRedux = __webpack_require__(/*! react-redux */ "./node_modules/react-redux/lib/index.js");
+
+var _ramda = __webpack_require__(/*! ramda */ "./node_modules/ramda/index.js");
+
+var _constants = __webpack_require__(/*! ./constants/constants */ "./src/constants/constants.js");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -33141,68 +33145,120 @@ var TreeContainer = function (_Component) {
     }, {
         key: 'render',
         value: function render() {
-            return _render(this.props.layout);
+            return recursivelyRender(this.props.layout, this.props.requestQueue);
         }
     }]);
 
     return TreeContainer;
 }(_react.Component);
 
-exports.default = TreeContainer;
-
-
 TreeContainer.propTypes = {
-    layout: _propTypes2.default.object
+    layout: _propTypes2.default.object,
+    requestQueue: _propTypes2.default.object
 };
 
-function _render(component) {
-    if (_ramda2.default.contains(_ramda2.default.type(component), ['String', 'Number', 'Null', 'Boolean'])) {
+function recursivelyRender(component, requestQueue) {
+    if ((0, _ramda.contains)((0, _ramda.type)(component), ['String', 'Number', 'Null', 'Boolean'])) {
         return component;
+    }
+
+    if ((0, _ramda.isEmpty)(component)) {
+        return null;
     }
 
     // Create list of child elements
     var children = void 0;
 
-    var componentProps = _ramda2.default.propOr({}, 'props', component);
+    var componentProps = (0, _ramda.propOr)({}, 'props', component);
 
-    if (!_ramda2.default.has('props', component) || !_ramda2.default.has('children', component.props) || typeof component.props.children === 'undefined') {
+    if (!(0, _ramda.has)('props', component) || !(0, _ramda.has)('children', component.props) || typeof component.props.children === 'undefined') {
         // No children
         children = [];
-    } else if (_ramda2.default.contains(_ramda2.default.type(component.props.children), ['String', 'Number', 'Null', 'Boolean'])) {
+    } else if ((0, _ramda.contains)((0, _ramda.type)(component.props.children), ['String', 'Number', 'Null', 'Boolean'])) {
         children = [component.props.children];
     } else {
         // One or multiple objects
         // Recursively render the tree
         // TODO - I think we should pass in `key` here.
-        children = (Array.isArray(componentProps.children) ? componentProps.children : [componentProps.children]).map(_render);
+        children = (Array.isArray(componentProps.children) ? componentProps.children : [componentProps.children]).map(function (child) {
+            return recursivelyRender(child, requestQueue);
+        });
     }
 
     if (!component.type) {
         /* eslint-disable no-console */
-        console.error(_ramda2.default.type(component), component);
+        console.error((0, _ramda.type)(component), component);
         /* eslint-enable no-console */
         throw new Error('component.type is undefined');
     }
     if (!component.namespace) {
         /* eslint-disable no-console */
-        console.error(_ramda2.default.type(component), component);
+        console.error((0, _ramda.type)(component), component);
         /* eslint-enable no-console */
         throw new Error('component.namespace is undefined');
     }
     var element = _registry2.default.resolve(component.type, component.namespace);
 
-    var parent = _react2.default.createElement.apply(_react2.default, [element, _ramda2.default.omit(['children'], component.props)].concat(_toConsumableArray(children)));
+    var parent = _react2.default.createElement.apply(_react2.default, [element, (0, _ramda.omit)(['children'], component.props)].concat(_toConsumableArray(children)));
+
+    // loading prop coming from TreeContainer
+    var isLoading = false;
+    var loadingProp = void 0;
+    var loadingComponent = void 0;
+
+    var id = componentProps.id;
+
+    if (requestQueue && requestQueue.filter) {
+        (0, _ramda.forEach)(function (r) {
+            var controllerId = (0, _ramda.isNil)(r.controllerId) ? '' : r.controllerId;
+            if (r.status === 'loading' && (0, _ramda.contains)(id, controllerId)) {
+                isLoading = true;
+
+                var _r$controllerId$split = r.controllerId.split('.');
+
+                var _r$controllerId$split2 = _slicedToArray(_r$controllerId$split, 2);
+
+                loadingComponent = _r$controllerId$split2[0];
+                loadingProp = _r$controllerId$split2[1];
+            }
+        }, requestQueue);
+
+        var thisRequest = requestQueue.filter(function (r) {
+            var controllerId = (0, _ramda.isNil)(r.controllerId) ? '' : r.controllerId;
+            return (0, _ramda.contains)(id, controllerId);
+        });
+        if (thisRequest.status === _constants.STATUS.OK) {
+            isLoading = false;
+        }
+    }
+
+    // Set loading state
+    var loading_state = {
+        is_loading: isLoading,
+        prop_name: loadingProp,
+        component_name: loadingComponent
+    };
 
     return _react2.default.createElement(
         _NotifyObservers2.default,
-        { key: componentProps.id, id: componentProps.id },
+        {
+            key: componentProps.id,
+            id: componentProps.id,
+            loading_state: loading_state
+        },
         parent
     );
 }
 
-_render.propTypes = {
-    children: _propTypes2.default.object
-};
+function mapStateToProps(state, ownProps) {
+    return {
+        layout: ownProps.layout,
+        loading: ownProps.loading,
+        requestQueue: state.requestQueue
+    };
+}
+
+exports.default = (0, _reactRedux.connect)(mapStateToProps)(TreeContainer);
 
 /***/ }),
 
@@ -34182,8 +34238,6 @@ Object.defineProperty(exports, "__esModule", {
 
 var _reactRedux = __webpack_require__(/*! react-redux */ "./node_modules/react-redux/lib/index.js");
 
-var _ramda = __webpack_require__(/*! ramda */ "./node_modules/ramda/index.js");
-
 var _actions = __webpack_require__(/*! ../../actions */ "./src/actions/index.js");
 
 var _react = __webpack_require__(/*! react */ "react");
@@ -34193,6 +34247,8 @@ var _react2 = _interopRequireDefault(_react);
 var _propTypes = __webpack_require__(/*! prop-types */ "./node_modules/prop-types/index.js");
 
 var _propTypes2 = _interopRequireDefault(_propTypes);
+
+var _ramda = __webpack_require__(/*! ramda */ "./node_modules/ramda/index.js");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -34220,6 +34276,8 @@ function mergeProps(stateProps, dispatchProps, ownProps) {
         children: ownProps.children,
         dependencies: stateProps.dependencies,
         paths: stateProps.paths,
+        loading_state: ownProps.loading_state,
+        requestQueue: stateProps.requestQueue,
 
         setProps: function setProps(newProps) {
             var payload = {
@@ -34242,7 +34300,8 @@ function NotifyObserversComponent(_ref) {
         id = _ref.id,
         paths = _ref.paths,
         dependencies = _ref.dependencies,
-        setProps = _ref.setProps;
+        setProps = _ref.setProps,
+        loading_state = _ref.loading_state;
 
     var thisComponentSharesState = dependencies && dependencies.find(function (dependency) {
         return dependency.inputs.find(function (input) {
@@ -34275,6 +34334,10 @@ function NotifyObserversComponent(_ref) {
         extraProps.setProps = setProps;
     }
 
+    if (children.props && !children.props.loading_state) {
+        extraProps.loading_state = loading_state;
+    }
+
     if (!(0, _ramda.isEmpty)(extraProps)) {
         return _react2.default.cloneElement(children, extraProps);
     }
@@ -34284,7 +34347,8 @@ function NotifyObserversComponent(_ref) {
 NotifyObserversComponent.propTypes = {
     id: _propTypes2.default.string.isRequired,
     children: _propTypes2.default.node.isRequired,
-    path: _propTypes2.default.array.isRequired
+    path: _propTypes2.default.array.isRequired,
+    loading_state: _propTypes2.default.object
 };
 
 exports.default = (0, _reactRedux.connect)(mapStateToProps, mapDispatchToProps, mergeProps)(NotifyObserversComponent);