+++ src/TreeContainer.js
@@ -1,46 +1,60 @@
 'use strict';
 
-import R from 'ramda';
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import Registry from './registry';
 import NotifyObservers from './components/core/NotifyObservers.react';
+import {connect} from 'react-redux';
+import {
+    isNil,
+    omit,
+    contains,
+    isEmpty,
+    forEach,
+    propOr,
+    type,
+    has,
+} from 'ramda';
+import {STATUS} from './constants/constants';
 
-export default class TreeContainer extends Component {
+class TreeContainer extends Component {
     shouldComponentUpdate(nextProps) {
         return nextProps.layout !== this.props.layout;
     }
 
     render() {
-        return render(this.props.layout);
+        return recursivelyRender(this.props.layout, this.props.requestQueue);
     }
 }
 
 TreeContainer.propTypes = {
     layout: PropTypes.object,
+    requestQueue: PropTypes.object,
 };
 
-function render(component) {
-    if (
-        R.contains(R.type(component), ['String', 'Number', 'Null', 'Boolean'])
-    ) {
+function recursivelyRender(component, requestQueue) {
+    if (contains(type(component), ['String', 'Number', 'Null', 'Boolean'])) {
         return component;
+    }
+
+    if (isEmpty(component)) {
+        return null;
     }
 
     // Create list of child elements
     let children;
 
-    const componentProps = R.propOr({}, 'props', component);
+    const componentProps = propOr({}, 'props', component);
 
     if (
-        !R.has('props', component) ||
-        !R.has('children', component.props) ||
+        !has('props', component) ||
+        !has('children', component.props) ||
         typeof component.props.children === 'undefined'
     ) {
         // No children
         children = [];
     } else if (
-        R.contains(R.type(component.props.children), [
+        contains(type(component.props.children), [
             'String',
             'Number',
             'Null',
@@ -55,18 +69,18 @@ function render(component) {
         children = (Array.isArray(componentProps.children)
             ? componentProps.children
             : [componentProps.children]
-        ).map(render);
+        ).map(child => recursivelyRender(child, requestQueue));
     }
 
     if (!component.type) {
         /* eslint-disable no-console */
-        console.error(R.type(component), component);
+        console.error(type(component), component);
         /* eslint-enable no-console */
         throw new Error('component.type is undefined');
     }
     if (!component.namespace) {
         /* eslint-disable no-console */
-        console.error(R.type(component), component);
+        console.error(type(component), component);
         /* eslint-enable no-console */
         throw new Error('component.namespace is undefined');
     }
@@ -74,13 +88,59 @@ function render(component) {
 
     const parent = React.createElement(
         element,
-        R.omit(['children'], component.props),
+        omit(['children'], component.props),
         ...children
     );
 
-    return <NotifyObservers key={componentProps.id} id={componentProps.id}>{parent}</NotifyObservers>;
+    // loading prop coming from TreeContainer
+    let isLoading = false;
+    let loadingProp;
+    let loadingComponent;
+
+    const id = componentProps.id;
+
+    if (requestQueue && requestQueue.filter) {
+        forEach(r => {
+            const controllerId = isNil(r.controllerId) ? '' : r.controllerId;
+            if (r.status === 'loading' && contains(id, controllerId)) {
+                isLoading = true;
+                [loadingComponent, loadingProp] = r.controllerId.split('.');
+            }
+        }, requestQueue);
+
+        const thisRequest = requestQueue.filter(r => {
+            const controllerId = isNil(r.controllerId) ? '' : r.controllerId;
+            return contains(id, controllerId);
+        });
+        if (thisRequest.status === STATUS.OK) {
+            isLoading = false;
+        }
+    }
+
+    // Set loading state
+    const loading_state = {
+        is_loading: isLoading,
+        prop_name: loadingProp,
+        component_name: loadingComponent,
+    };
+
+    return (
+        <NotifyObservers
+            key={componentProps.id}
+            id={componentProps.id}
+            loading_state={loading_state}
+        >
+            {parent}
+        </NotifyObservers>
+    );
 }
 
-render.propTypes = {
-    children: PropTypes.object,
-};
+function mapStateToProps(state, ownProps) {
+    return {
+        layout: ownProps.layout,
+        loading: ownProps.loading,
+        requestQueue: state.requestQueue,
+    };
+}
+
+export default connect(mapStateToProps)(TreeContainer);